requirements.txt
celery==5.2.3
click==8.1.3
Flask==2.1.2
Flask-WTF==1.0.0
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
redis==4.1.4
Werkzeug==2.1.2
zipp==3.8.0
python-dotenv~=0.20.0